fix(mediatask): rename to snapshotQueue
Rename snapshotSubtitleQueue to snapshotQueue.
