+++ src/renderer/plugins/mediaTasks/snapshotQueue.ts
